补充辅助函数，经过调试找到一个 bug，central_directory_file_header 里面的 external_file_attrs 应该是 uint32_t，bug 的版本写成了 uint16_t

+++ source/runtime/pnnx/utils/storezip.cpp
@@ -53,7 +53,7 @@ namespace cnnx
         uint16_t file_comment_length;
         uint16_t start_disk;
         uint16_t internal_file_attrs;
-        uint16_t external_file_attrs;
+        uint32_t external_file_attrs;
         uint32_t lfh_offset;
         });
